Save the current buffer to `ezbl-instance'
Since `ezbl-start' should be called with the buffer embedding the Ezbl window as
the current buffer, save that buffer to the instance. This allows consistent
access to the buffer, even if its name changes. This is important because the
buffer name will be set to track the title of the page it is embedding.

Signed-off-by: Dan Hackney <dan@haxney.org>