Eliminate usage of a "name" string for instances.
Since the uri and title of a page can change, it does not make sense to assign
any long-lasting string identifier to a particular ezbl instance. Use the pid
from the process for identifying purposes instead.

Also, make the distinction between the "display-buffer" and the "output-buffer"
more clear. The display-buffer is the one with the embedded xwidget in it, and
the output-buffer is the buffer which receives the STDOUT and STDERR from the
Uzbl process.

Signed-off-by: Dan Hackney <dan@haxney.org>