Rename the display buffer to include the Uzbl pid.
Makes the display buffers have a consistent naming system, but makes it more
awkward to select an Ezbl buffer by its name, since the buffer name would be
something like "*ezbl-display-12345*". An alternate buffer selection system
could be used, or this change could be reverted, if it is too cumbersome.

Signed-off-by: Dan Hackney <dan@haxney.org>