Correct `ezbl-get-inst' when given an instance.
Under the old behavior of `ezbl-instance-p', it would return its argument if it
was an instance, and `nil' if not. The `ezbl-inst-p' created by `defstruct'
returns `t' if the argument is an instance. This means that using the return
value of `ezbl-inst-p' will always be either `nil' or `t', so inst must be
separately passed along.

Signed-off-by: Dan Hackney <dan@haxney.org>