Fix unmatched event filter handling.
Events that were actually being handled were being printed to the process output
buffer, due to the fact that `answer', rather than `ans' was being printed. This
meant that if the filter received multiple lines at once (which is fairly
common) and if just one of those lines did not match the regular expression,
then the whole batch of lines would be printed, rather than just those that
failed. This lead to a great deal of spurious output.

Signed-off-by: Dan Hackney <dan@haxney.org>