Make `ezbl-inst' slots `setf'-able.
Since `setf' handles arguments differently than regular function applications,
the advice on the slot accessors was not sufficient to allow `setf' to resolve
`ezbl-inst' structs.

When invoked as `(setf (ezbl-inst-vars 12345) new-vars)', (assuming `12345' is a
valid pid for an `ezbl-inst') `setf` didn't know that it should call
`ezbl-inst-get' to resolve `12345' to the corresponding `ezbl-inst'.

The `setf-method' of the slot accessors is redefined to perform this resolution
(in strict mode, signaling an error if resolution is not possible) before
performing the rest of the assignment.

Signed-off-by: Dan Hackney <dan@haxney.org>